requirements.txt
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
expat=2.6.3=h6a678d5_0
filelock=3.13.1=pypi_0
fonttools=4.54.1=pypi_0
fsspec=2024.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
grpcio=1.67.0=pypi_0
idna=3.10=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py312h06a4308_0
protobuf=5.28.2=pypi_0
psutil=6.1.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyparsing=3.2.0=pypi_0
python=3.12.1=h996f2a0_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-triton-rocm=3.1.0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.1.0=py312h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
torch=2.5.0+rocm6.2=pypi_0
torch-tb-profiler=0.4.3=pypi_0
torchaudio=2.5.0+rocm6.2=pypi_0
torchvision=0.20.0+rocm6.2=pypi_0
tqdm=4.66.5=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024.2=pypi_0
ultralytics=8.3.18=pypi_0
ultralytics-thop=2.0.9=pypi_0
urllib3=2.2.3=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.44.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
